display top files

+++ Files.cs
@@ -5,7 +5,7 @@ namespace topfiles
     public class Files{
         public string filename {get; set;}
         public DateTime time {get; set; }
-        public int filesize {get; set; }
+        public long filesize {get; set; }
 
         public Files(string file){
             filename = file;
